some little improvement
--HG--
extra : convert_revision : svn%3A2597059e-4560-0410-9562-ba342e98936d/trunk%40295

+++ modules/hfnuadmin/controllers/category.classic.php
@@ -121,12 +121,15 @@ class categoryCtrl extends jController {
 		if ($id_cat == 0) {
 			jMessage::add(jLocale::get('hfnuadmin~category.invalid.datas'),'error');
 		} else {		
-			//@TODO : check if there is existing Forum linked to this category
-			// if so, ask the user if he wants to remove the forum and all the linked posts !
-			$dao = jDao::get('havefnubb~category');        
-			$dao->delete($id_cat);
-        
-	        jMessage::add(jLocale::get('hfnuadmin~category.category.deleted'),'ok');
+			$rec = jDao::get('havefnubb~forum_cat')->countByIdCat($id_cat);
+			if ($rec == 0) {
+				$dao = jDao::get('havefnubb~category');        
+				$dao->delete($id_cat);        
+				jMessage::add(jLocale::get('hfnuadmin~category.category.deleted'),'ok');
+			}
+			else {
+				jMessage::add(jLocale::get('hfnuadmin~category.category.cant.be.deleted',$rec),'error');
+			}
 		}
 		
         $rep = $this->getResponse('redirect');
